+++ client/src/test/data/concept/imprints.ts
@@ -1,0 +1,61 @@
+export const imprints = {
+  id: 'https://lux.collections.yale.edu/data/concept/1a05484c-59d1-4e65-be9b-8e54c5bd18c6',
+  type: 'Type',
+  equivalent: [
+    {
+      id: 'http://id.loc.gov/authorities/subjects/sh99001566',
+      type: 'Type',
+    },
+    {
+      id: 'http://www.wikidata.org/entity/Q2608849',
+      type: 'Type',
+    },
+  ],
+  identified_by: [
+    {
+      type: 'Name',
+      content: 'Imprints',
+      language: [
+        {
+          id: 'https://lux.collections.yale.edu/data/concept/dfa53b96-4eda-4c9a-b091-10008a726c38',
+          type: 'Language',
+          _label: 'English',
+          notation: 'en',
+          equivalent: [
+            {
+              id: 'http://vocab.getty.edu/aat/300388277',
+              type: 'Language',
+              _label: 'English',
+            },
+          ],
+        },
+      ],
+      classified_as: [
+        {
+          id: 'https://lux.collections.yale.edu/data/concept/f7ef5bb4-e7fb-443d-9c6b-371a23e717ec',
+          type: 'Type',
+          _label: 'Primary Name',
+          equivalent: [
+            {
+              id: 'http://vocab.getty.edu/aat/300404670',
+              type: 'Type',
+              _label: 'Primary Name',
+            },
+          ],
+        },
+        {
+          id: 'https://lux.collections.yale.edu/data/concept/ab99d278-9323-4d84-8e97-1846058fc587',
+          type: 'Type',
+          _label: 'Sort Name',
+          equivalent: [
+            {
+              id: 'http://vocab.getty.edu/aat/300451544',
+              type: 'Type',
+              _label: 'Final: Sort Title',
+            },
+          ],
+        },
+      ],
+    },
+  ],
+}